+++ gradle-plugin/src/main/java/com/deezer/caupain/plugin/DependencyUpdatePlugin.kt
@@ -64,6 +64,7 @@ open class DependencyUpdatePlugin : Plugin<Project> {
             pluginRepositories.convention(ext.repositories.plugins)
             useCache.convention(ext.useCache)
             onlyCheckStaticVersions.convention(ext.onlyCheckStaticVersions)
+            gradleStabilityLevel.convention(ext.gradleStabilityLevel)
         }
     }